fix(RadioButton): RadioButton looks wrong when zoming in browser
The issue is caused by rounding errors for the inner checkmark
when the outer box uses a border that gets scaled by small fractions of
a px.

Fixed by using outline instead of border

Also changed the checkmark from background-collor to box-shaddow
so that it is visible in print mode. It's also way less likely to get
roundoff errors when drawing a shaddow than when

I also updated the storybook to show all the relevant variants in checked mode
(missing ones was compact, error, and disabled)