feat(Avatar): new component (#2312)
resolves #1098

In contrast to `Badge`s `overlap`, we use `square` instead of
`rectangle`. This is because the avatar is **always** a square.
The component does not support `asChild`, since elements such as
`button` or `a` should wrap the element. Styling is up to the consumers
to handle.
Font class will be removed later in favor of `composes` in modules

---------

Co-authored-by: Michael Marszalek <mimarz@gmail.com>
Co-authored-by: Eirik Backer <eirik.backer@gmail.com>