refactor(cli): add and use getThemesFor(dimensions) in build command
This automates the config filtering which so far has been done manually, e.g. this
```ts
const semanticThemes = R.filter((val) => val.mode === 'light' && val.typography === 'primary', themes)
```

is now this
```
const semanticThemes = getThemesFor('semantic');
```

The main reason for doing this is that it requires no changes to the filters when adding new modes, while the code as it is before this commit would require every `R.filter(...)` statement to be updated to ignore irrelevant modes -- or suffer exponential increase in build time.

The resulting css files from the build command remain unchanged.