feat(cli,react): generate data-color typings based on tokens (#2795)
**Most importantly**, `data-color` now has type safety based on the
token structure which is used to generate the theme CSS. Instructions
for enabling this for consumers has been added. An optional
`react-types.d.ts` has also been added, which adds type hints for
`data-color` and `data-size` to all HTML elements.

Closes #2655 

## Other changes

### tsconfig restructuring

Restructures the React library's TypeScript config into test, stories &
lib-specific configs, with a common `tsconfig.json` that just references
all three. This allows individual settings, e.g. only including vitest
types in the `.test.ts*` files or only allowing ALL colors in
`.stories.ts*` (`tsconfig.lib.json` only allows colors that we know
exist regardless the theme in use).

This gives the following structure under `packages/react`, with each
sub-config defining a composite project which supports incremental
compilation:
```mermaid
graph LR;
    tsconfig.json --> tsconfig.lib.json;
    tsconfig.json --> tsconfig.stories.json;
    tsconfig.json --> tsconfig.tests.json;
    tsconfig.stories.json --> tsconfig.lib.json;
    tsconfig.tests.json --> tsconfig.lib.json;
```

The updated config also fixes a weird issue with interface declaration
merging which didn't work. Not sure why, but it was fixed by outputting
`.js` and `.d.ts` files from `tsc` to the same directory (which is the
default). This complicates the Rollup config a tiny bit, but not much.

The root `tsconfig.json` has been renamed to `tsconfig.base.json` to
make it clear that it is not an actual project config, but only a base
config.

### vitest workspace config

Test config specific to `@digdir/designsystemet-react` has been moved to
`packages/react/vitest.config.mjs`, and the root workspace config has
been updated to look for projects under `packages/*`. This is good
practice in a monorepo, allowing us to have test config specific to
projects, e.g. if we want to add unit tests to `cli` which don't need
`jsdom`.

This was also necessary to allow vitest to resolve the correct
`tsconfig.tests.json` file for the React library.