fix: :bug: Fix missing selector on "default" color
Found a bug when building that first color did not get `:root, [data-color-scheme],` selector. This was due to `buildOptions` not being defined. Fixing this for now, but we should rewrite this to avoid hidden globals. All functions should be pure functions, so that its clear for maintaners where dependencies/parameters come from and avoid side-effects.