chore: rename sizing & spacing CSS variables to size (#2935)
Resolves #2934 

Made a shim for renaming `spacing` & `sizing` to `size` as design-tokens
for theme will be done in a separate PR.

Added shim in `css/src/index.css`, apart from that just search and
replace `spacing`/`sizing` with `size`.

Did not update all code under Figma plugin as there are som duplicates
there that needs to be resolved first.