style: only apply translate if outside `.ds-field` (#2986)
Reported on slack.


![image](https://github.com/user-attachments/assets/69ce063f-9941-46c8-aba6-e0cea2c4ca17)

This does **not** fix the issue above. If the consumer changes `Field`s
display, i.e. `grid` to `flex`, and then `align-items: center`, we can't
ensure the input is aligned to the first line of the label anymore. In
these cases, it is excpected that the consumer needs to make more css
changes, in this case remove our `translate`.

This PR does make sure a standalone checkbox or radio does not have
`translate`