chore(css): go over and make consistent, highlight issues (#3002)
#2638

## Highlights:
- Sometimes we use a var `-color`, and then use it for `background:`. In
these cases it is to add color to an icon - so the usage is correct.
`badge.css` is an example of this.
- We use the word `spacing` in our tokens quite a bit. This is reserved
for where we use it to do more than just set a property. For example in
padding, but also in a calc for `margin-inline`
- We were incosistent in placement of css props on the root classname. I
moved this to just after the initial css var definitions.
- Where we have used vars for icons, all of them have been changed to
use `-icon`. `Details` had `chevron`, for example. `Search` is an
exception to this, as it has multiple icons. Both of them are named in
this case,
- We only use `background`, not `background-color`
- `border:` now uses two properties. The color is a variable, and the
width and style is a css prop, unless used multiple places in the same
component. `border` and `border-color` will still stay as separate vars

**`combobox.css` has not been touched**