chore: fix wrong color name returned from CLI func (#3082)
Add the base prefix to the contrast color names in the
`getColorInfoFromPosition` function.

In the themebuilder the preview examples loop through the function to
set unique tokens for each card. Now the contrast colors are not applied
because the base prefix is missing. When a very light color is selected
in the themebuilder, the text in the Button on the preview Cards are
white.